+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Controllers/EventAdminController.cs
@@ -174,23 +174,21 @@ namespace AllReady.Areas.Admin.Controllers
 
             if (ModelState.IsValid)
             {
-                var existingImageUrl = campaignEvent.ImageUrl;
-
                 if (fileUpload != null)
                 {
                     if (fileUpload.IsAcceptableImageContentType())
                     {
+                        var existingImageUrl = campaignEvent.ImageUrl;
                         campaignEvent.ImageUrl = await _imageService.UploadEventImageAsync(campaign.OrganizationId, campaignEvent.Id, fileUpload);
+                        if (campaignEvent.ImageUrl != null && existingImageUrl != null)
+                        {
+                            await _imageService.DeleteImageAsync(existingImageUrl);
+                        }
                     }
                     else
                     {
                         ModelState.AddModelError("ImageUrl", "You must upload a valid image file for the logo (.jpg, .png, .gif)");
                         return View(campaignEvent);
-                    }
-
-                    if (existingImageUrl != null)
-                    {
-                        await _imageService.DeleteImageAsync(existingImageUrl);
                     }
                 }
